Script refactoring and TA3 testbed integration (#93)
Script refactoring and TA3 testbed integration
==================================

This PR implements the following.

General script refactoring
--------------------------

- The scripts under the `tools` directory no longer have extensions in their filenames (`.sh`, `.scpt`, etc). One advantage of this is that end-users should be able to just double-click an executable script rather than relying on the terminal.
- A bit more structure has been imposed on the `tools` directory - there is now a hierarchical structure for the scripts, with scripts grouped logically into directories.
- Keyboard interrupts (SIGINTs) are trapped now in run_session - they completely kill Minecraft and `ffmpeg` processes and perform appropriate cleanup.
- Parts of the `run_session` script have been pulled out into another script, `configure_session`. The configure_session script detects the operating system version and appropriately sets a number of environment variables, thus consolidating the large number of if-else statements that were previously in the `run_session` script.
- We can now add `set -u` at the top of the scripts without issue, since there is a bash syntax feature that allows for specifying default values for variables if they have not yet been defined. For example, `"${VARIABLE:-DEFAULT}"` means that the variable `VARIABLE` is assigned the default value `DEFAULT` if it has not been previously assigned. So we can now easily set environment variables at the time of invocation. Variables relevant to ToMCAT devs include `ENABLE_MQTT`, `ENABLE_FFMPEG`, `do_tutorial_mission` (you may often want to set this to 0 to skip the tutorial mission). This syntax feature has been tested on both macOS Catalina and the Github Actions Ubuntu 18.04 VM runner.
- On macOS, the scripts check whether the macOS version is at least 10.15 (Catalina) - if not, then system audio recording is disabled.


TA3 testbed integration
------------------------

### MQTT message bus publishing

Support has been added for publishing observations to an MQTT message bus using the `mosquitto` message broker. This feature is optional and disabled by default, but can be enabled using by setting the `ENABLE_MQTT` environment variable to 1. The current version of the script expects that the `mosquitto` broker is up and running if `ENABLE_MQTT` is set to 1, but a future version of the script will check this explicitly, and start the broker if it is not already up.

The observations are produced by the `runExperiment` executable. We no longer pass the `--record_path` argument to the executable, but instead redirect the output either to 
1. a text file called `malmo_data.txt` (if `ENABLE_MQTT=0` (the default) (instead of the previous gzipped directory `malmo_data.tgz`), or
2. the `mosquitto_pub` executable, which takes in the JSON messages and publishes them to the message bus (the default host (`localhost`) and port (1883) are used - perhaps this might change in the future). This option is activated when `ENABLE_MQTT=1`

The messages are output to the `observations/state` topic. A caveat: the messages currently are close to the JSON schema used by the TA3 testbed, but are not quite exactly the same format (I wanted to get this branch merged in before it became too long-lived).

#### Notes for ToMCAT team members

Mission dev group members: To run a session using the message broker, first start up the `mosquitto` broker in the background with
 
    mosquitto &

Then run 

    ENABLE_MQTT=1 ./tools/run_session

In a separate terminal window, you can follow the messages being published with the command

    mosquitto_sub -t observations/state

Finally, you can optionally shut down the mosquitto broker with

    pkill mosquitto

@Adi-UA @paulosoaresua If you can implement streaming JSON output objects to a file, then we can publish those to the message bus - we would either publish all events to a single topic (`observations/events`) or divide them among two topics (`observations/events/playeraction` and `observations/events/entityaction`)

@paulosoaresua @kobus-barnard A few notes on agent design, mostly to help myself remember later on:

The ToMCAT agent should be a executable (separate from `runExperiment`) that reads JSON in line-by-line (i.e. takes in an incoming stream of messages) and processes them. If the executable is named something like `tomcat_agent`, then the invocation might be something like

    mosquitto_sub -t observations/state | ./tomcat_agent

The JSON being output in `Mission.cpp` uses the `nlohmann-json` C++ library (https://github.com/nlohmann/json), which is quite easy to use, so I this we should suggest using that in the `tomcat_agent` executable as well. An exercise to test minimal connectivity would be to just have the agent do a roundtrip parse of the incoming JSON - i.e. parse each incoming JSON string into a `nlohmann::json` object using `json::parse`, perform a nominal modification to the object, and dump it to standard output using the `dump()` method. We can then publish the modified objects to the message bus under the `agent/tom` topic. The complete invocation for this process might be something like 

    mosquitto_sub -t observations/state | ./tomcat_agent | mosquitto_pub -t agent/tom

(Of course I haven't tested the above approach, but if that is the case it would be a pretty elegant and easy way to integrate with the TA3 testbed without having to actually link `tomcat_agent` against `libmosquitto`.)

### Support (work in progress) for USAR Singleplayer mission

<img width="966" alt="Screen Shot 2020-04-12 at 5 29 04 PM" src="https://user-images.githubusercontent.com/653549/79160029-f60b0700-7d8d-11ea-87d0-40ffe6491d6b.png">

Partial support has been added to run TA3 mission worlds in our framework. So far, I've tested that we can load the world in our framework and set the start position using a Java class that inherits from `Mission.java`. Potentially we will be able to easily have self-reports etc using the collapsed office building environment as well. 

Two new mission classes, `USARSinglePlayerMission` and `USARSinglePlayerClientMission` have been added (mission ID 2) - if the appropriate world directory (`Singleplayer`) is present under the `tomcat/data/worlds` directory. ToMCAT project members can download the world from the ASIST-ToMCAT shared Google Drive folder (under the `Data/` directory) ([link](https://drive.google.com/open?id=14P9APdiv6PQRIhOilP2IQFv_aaq1V5Sa))

- All instances of SAR have been replaced by 'Zombie' - this is to differentiate our mission from the urban search-and-rescue (USAR) mission being developed by Aptima+ASU, which this PR implements support for.
- The `sar` folder in `worlds.tgz` has been renamed `zombie`, and a couple of new dependencies (`nlohmann-json`, `mosquitto`, and `switchaudio-osx` (macOS-only)) have been added in `tools/install_dependencies`) so once this PR is merged, people should run `./tools/update_tomcat` to update their dependencies and get the new worlds folder.

To run the USAR Singleplayer mission you can invoke the following command (provided you have put the `Singleplayer` directory under `tomcat/data/worlds`:

```
mission=2 ./tools/run_session
```

Miscellaneous
--------------

- `docs/installation.md` has been updated to correctly reflect the number of files that will be produced.
- `src/Mission.cpp` now checks to see if the requested ToMCAT world folder exists and throws an exception if not.
- `src/Python_wrapper.cpp` has been removed to reduce maintenance burden.